requirements.txt
streamlit
numpy
matplotlib
torch
torchvision
pillow
requests
pandas
PyPDF2
nltk